darwin: adjust syscall suffix for arm64
